+++ python/GafferSceneUI/RenderPassEditor.py
@@ -126,6 +126,7 @@ class RenderPassEditor( GafferSceneUI.SceneEditor ) :
 			# for header-only context menus in CatalogueUI.
 			self.__pathListing._qtWidget().header().setContextMenuPolicy( QtCore.Qt.CustomContextMenu )
 			self.__pathListing._qtWidget().header().customContextMenuRequested.connect( Gaffer.WeakMethod( self.__headerContextMenuRequested ) )
+			self.__pathListing._qtWidget().header().sectionPressed.connect( Gaffer.WeakMethod( self.__sectionPressed ) )
 			self.__pathListing._qtWidget().header().sectionMoved.connect( Gaffer.WeakMethod( self.__sectionMoved ) )
 
 			with GafferUI.ListContainer( GafferUI.ListContainer.Orientation.Horizontal, spacing = 4 ) :
@@ -154,6 +155,7 @@ class RenderPassEditor( GafferSceneUI.SceneEditor ) :
 
 		self.__columnCache = {}
 		self.__ignoreSectionMoved = False
+		self.__stretchedFavouriteColumn = None
 
 		self._updateFromSet()
 		self.__updateColumns()
@@ -319,7 +321,17 @@ class RenderPassEditor( GafferSceneUI.SceneEditor ) :
 			if len( sectionColumns ) == 0 :
 				# Include a blank spacer column when there are no favourites.
 				sectionColumns.append( ( GafferUI.StandardPathColumn( "", "", GafferUI.PathColumn.SizeMode.Stretch ), 0 ) )
+			elif self.__stretchedFavouriteColumn != sectionColumns[-1][0] :
+				# We don't want our _AdderColumn to automatically stretch,
+				# so stretch the second-to-last column instead, resetting any
+				# previously stretched column in the process.
+				if self.__stretchedFavouriteColumn is not None :
+					self.__stretchedFavouriteColumn.setSizeMode( GafferUI.PathColumn.SizeMode.Default )
 
+				self.__stretchedFavouriteColumn = sectionColumns[-1][0]
+				self.__stretchedFavouriteColumn.setSizeMode( GafferUI.PathColumn.SizeMode.Stretch )
+
+			sectionColumns.append( ( self.__acquireColumn( _AdderColumn, currentSection ), -1 ) )
 		else :
 			for groupKey, sections in self.__columnRegistry.items() :
 				if IECore.StringAlgo.match( tabGroup, groupKey ) :
@@ -393,9 +405,42 @@ class RenderPassEditor( GafferSceneUI.SceneEditor ) :
 
 		return self.settings()["section"].getValue() == "Favourites"
 
+	def __showOptionMenu( self ) :
+
+		m = IECore.MenuDefinition()
+
+		favourites = self.settings()["favouriteColumns"].getValue()
+
+		enabledOptionPrefixes = set( [ "render", "renderPass", "sampleMotion" ] )
+		for renderer in Gaffer.Metadata.targetsWithMetadata( "renderer:*", "optionPrefix" ) :
+			if Gaffer.Metadata.value( renderer, "ui:enabled" ) is not False :
+				enabledOptionPrefixes.add( Gaffer.Metadata.value( renderer, "optionPrefix" ).rstrip( ":" ) )
+
+		for option in Gaffer.Metadata.targetsWithMetadata( "option:*", "defaultValue" ) :
+			if option.split( ":" )[1] not in enabledOptionPrefixes :
+				continue
+
+			category = Gaffer.Metadata.value( option, "category" ) or "Other"
+			section = ( Gaffer.Metadata.value( option, "layout:section" ) or "Other" ).replace( ".", "/" )
+			label = Gaffer.Metadata.value( option, "label" ) or option[7:]
+			m.append(
+				f"/{category}/{section}/{label}",
+				{
+					"command" : functools.partial( Gaffer.WeakMethod( self.__favourite ), option ),
+					"active" : option not in favourites,
+				}
+			)
+
+		self.__contextMenu = GafferUI.Menu( m )
+		self.__contextMenu.popup( parent = self )
+
 	def __headerContextMenuRequested( self, pos ) :
 
 		column = self.__pathListing.columnAt( imath.V2f( pos.x(), pos.y() ) )
+		if isinstance( column, _AdderColumn ) :
+			self.__showOptionMenu()
+			return
+
 		m = IECore.MenuDefinition()
 		userEditableSection = self.__currentSectionEditable()
 		if isinstance( column, GafferSceneUI.Private.InspectorColumn ) :
@@ -458,10 +503,19 @@ class RenderPassEditor( GafferSceneUI.SceneEditor ) :
 
 		firstMovableIndex = len( self.__commonColumns )
 		favourites = list( self.settings()["favouriteColumns"].getValue() )
-		favourites.insert( max( 0, newVisualIndex - firstMovableIndex ), favourites.pop( oldVisualIndex - firstMovableIndex ) )
+		if len( favourites ) > oldVisualIndex - firstMovableIndex :
+			favourites.insert( max( 0, newVisualIndex - firstMovableIndex ), favourites.pop( oldVisualIndex - firstMovableIndex ) )
 
 		self.settings()["favouriteColumns"].setValue( IECore.StringVectorData( favourites ) )
 		self.__resetSectionOrder()
+
+	def __sectionPressed( self, logicalIndex ) :
+
+		if isinstance( self.__pathListing.getColumns()[logicalIndex], _AdderColumn ) :
+			self.__showOptionMenu()
+			return True
+
+		return False
 
 	def __displayGroupedChanged( self ) :
 
@@ -857,6 +911,20 @@ class RenderPassEditor( GafferSceneUI.SceneEditor ) :
 		self.__addButton.setToolTip( "Click to add render pass." if editable else "To add a render pass, first choose an editable Edit Scope." )
 
 GafferUI.Editor.registerType( "RenderPassEditor", RenderPassEditor )
+
+class _AdderColumn( GafferUI.PathColumn ) :
+
+	def __init__( self, *args ) :
+
+		GafferUI.PathColumn.__init__( self )
+
+	def cellData( self, path, canceller ) :
+
+		return GafferUI.PathColumn.CellData( value = "", toolTip = "Click on the header to add columns." )
+
+	def headerData( self, canceller ) :
+
+		return GafferUI.PathColumn.CellData( value = "", icon = "plus.png", toolTip = "Click to add columns." )
 
 ##########################################################################
 # Metadata controlling the settings UI